+++ lib/screens/genre_screen.dart
@@ -1,4 +1,4 @@
-import 'package:finamp/components/GenreScreen/genre_screen_provider.dart';
+import 'package:finamp/services/genre_screen_provider.dart';
 import 'package:finamp/services/finamp_user_helper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
